Finished validation logic and added a modal
Finished the rest of the validation logic
for the rest of the fields and made sure that
the confirm password triggers whenever
the user changes the original password,
or when the user is typing in the confirm
password field, or if the user even focuses on
the field to display the message with
.reportValidity(). Also prevented submission
by adding an event listener to the
submit event and prevent default and reuturned
if the .reportValidity was false. .reportValidity
also displays any error messages that are still
active.